+++ packages/atomic-hosted-page/src/atomic-hosted-page.esm.ts
@@ -1,0 +1,3 @@
+import './autoloader/index.js';
+
+export * from './index.js';