+++ packages/headless/src/api/service/insight/get-interface/get-interface-config-response.ts
@@ -11,6 +11,7 @@ export interface InsightInterface {
   facets: Facet[];
   tabs: Tab[];
   settings: SettingsSection;
+  answerConfigId?: string;
 }
 
 interface InsightResultTemplate {
@@ -82,10 +83,14 @@ interface Tab {
   conditions: Condition[];
 }
 
+interface InsightGenQAOptions extends InsightOption {
+  collapsible: boolean;
+}
+
 interface SettingsSection {
   createArticle: InsightOption;
   fullSearch: InsightOption;
-  genQA: InsightOption;
+  genQA: InsightGenQAOptions;
   userActions: InsightUserActionOptions;
 }